backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
sounddevice>=0.4.6
numpy>=1.24.0
websockets>=12.0
openai>=1.6.0
Pillow>=10.0.0
pynput>=1.7.6
python-dotenv>=1.0.0
janus>=1.0.0
